+++ main/src/test/java/net/oneandone/stool/StoolIT.java
@@ -80,7 +80,7 @@ public class StoolIT {
         stool("status", "-stage", "it");
         stool("validate", "-stage", "it");
         stool("config", "-stage", "it", "memory=300");
-        stool("build", "-v", "-stage", "it");
+        stool("build", "-v", project.getAbsolute());
         stool("start", "-v", "-stage", "it");
         stool("validate", "-stage", "it");
         stool("status", "-stage", "it");